12月悦来需求优化：活动有效期 增加 分钟

+++ src/main/java/com/xiliulou/electricity/query/ShareMoneyActivityAddAndUpdateQuery.java
@@ -35,7 +35,7 @@ public class ShareMoneyActivityAddAndUpdateQuery {
     /**
      * 有效时间，单位：分钟
      */
-    private Long minutes;
+    private Integer minutes;
     
     /**
      * 金额
